added a policy authotization to updating and deleting job listing

+++ app/Http/Controllers/JobController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Job;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -11,6 +12,8 @@ use Illuminate\View\View;
 
 class JobController extends Controller
 {   
+    use AuthorizesRequests;
+
     // @desc Show all job listing
     // @route GET /jobs    
     public function index()
@@ -80,6 +83,7 @@ class JobController extends Controller
     // @route GET /jobs/{$id}/edit  
     public function edit(Job $job): View
     {
+        $this->authorize("update", $job);
         return view("jobs.edit")->with("job", $job);
     }
 
@@ -87,6 +91,9 @@ class JobController extends Controller
     // @route PUT /jobs/{$id}
     public function update(Request $request, Job $job)
     {
+        //Check if user is authorized
+        $this->authorize("update", $job);
+
          $validatedData = $request->validate([
             "title" => "required|string|max:255",
             "description" => "required|string",
@@ -133,6 +140,8 @@ class JobController extends Controller
     // @route DELETE /jobs/{$id}  
     public function destroy(Job $job): RedirectResponse
     {
+        $this->authorize("delete", $job);
+
         // If logo, then delete it
         if ($job->company_logo && Storage::disk('public')->exists($job->company_logo)) {
             Storage::disk('public')->delete($job->company_logo);
